+++ addon/globalPlugins/nvda_json/abstract.py
@@ -7,6 +7,8 @@ import ui
 
 import wx
 
+from .gui_components import EVT_AUTOCOMPLETE_DELETE, EVT_AUTOCOMPLETE_ENTER, AutoCompleteTextCtrl
+
 SAVED_EXPRESSIONS_FILE_NAME = os.path.dirname(__file__)+'/../../../../json-expressions.json'
 
 class JsonManipulatorDialog(wx.Dialog):
@@ -34,7 +36,7 @@ class JsonManipulatorDialog(wx.Dialog):
         json_sizer.Add(label_original_text)
         json_sizer.Add(self.original_text, 1, wx.EXPAND | wx.ALL, 5)
         self.label_manipulation_expression = wx.StaticText(self)
-        self.manipulation_expression = wx.TextCtrl(self, style=wx.TE_LEFT|wx.TE_PROCESS_ENTER )
+        self.manipulation_expression = AutoCompleteTextCtrl(self, choices = self.get_auto_complete_choices(), style=wx.TE_LEFT)
         json_sizer.Add(self.label_manipulation_expression)
         json_sizer.Add(self.manipulation_expression, 0, wx.EXPAND | wx.ALL, 5)
         label_output = wx.StaticText(self, label='Output')
@@ -44,11 +46,13 @@ class JsonManipulatorDialog(wx.Dialog):
         copy_output_button = wx.Button(self, label='Copy output to clipboard')
         copy_output_button.Bind(wx.EVT_BUTTON, self.on_copy_output_click)
         json_sizer.Add(copy_output_button, 0, wx.ALIGN_RIGHT | wx.ALL, 5)
-        main_sizer.Add(json_sizer, 1, wx.EXPAND | wx.ALL, 5)
+        self.json_sizer = json_sizer
+        main_sizer.Add(self.json_sizer, 1, wx.EXPAND | wx.ALL, 5)
         self.SetSizer(main_sizer)
         self.Bind(wx.EVT_CHAR_HOOK, self.on_key)
-        self.manipulation_expression.Bind(wx.EVT_TEXT_ENTER, self.manipulate)
-        self.manipulation_expression.Bind(wx.EVT_KEY_DOWN, self.on_manipulation_expression_key_down)
+        self.manipulation_expression.Bind(EVT_AUTOCOMPLETE_ENTER, self.on_manipulation_expression_enter)
+        self.manipulation_expression.Bind(EVT_AUTOCOMPLETE_DELETE, self.on_manipulation_expression_delete)
+        self.manipulation_expression.Bind(wx.EVT_CHAR_HOOK, self.on_manipulation_expression_key_down)
 
     def on_copy_output_click(self, event):
         copied = api.copyToClip(self.output.GetValue())
@@ -65,14 +69,31 @@ class JsonManipulatorDialog(wx.Dialog):
 
     def on_manipulation_expression_key_down(self, event):
         key_code = event.GetKeyCode()
-        if key_code == wx.WXK_RETURN and event.ControlDown():
+        if key_code == wx.WXK_RETURN:
             try:
                 self.manipulate(event)
+                self.output.SetFocus()
             except Exception as e:
                 return
-            self.save_expression()
+            if event.ControlDown():
+                self.save_expression()
+                self.manipulation_expression.set_choices(self.get_auto_complete_choices())
         else:
             event.Skip()
+
+    def on_manipulation_expression_enter(self, event):
+        self.manipulate(event)
+        self.output.SetFocus()
+
+    def on_manipulation_expression_delete(self, event):
+        expression = event.GetValue()
+        self.saved_expressions = [
+        item for item in self.saved_expressions
+        if not (item['expression'] == expression and item['type'] == self.expression_type)
+        ]
+        self.save_expressions_file()
+        self.manipulation_expression.set_choices(self.get_auto_complete_choices())
+        ui.message(f"Query '{expression}' removed!")
 
     def save_expression(self):
         expression = self.manipulation_expression.GetValue().strip()
@@ -80,8 +101,8 @@ class JsonManipulatorDialog(wx.Dialog):
             ui.message('Empty expression')
             return
         expression_exists = any(
-            saved['type'] == self.expression_type and saved['expression'] == expression
-            for saved in self.saved_expressions
+            item['type'] == self.expression_type and item['expression'] == expression
+            for item in self.saved_expressions
         )
         if expression_exists:
             ui.message('Query already exists')
@@ -90,12 +111,22 @@ class JsonManipulatorDialog(wx.Dialog):
             'type': self.expression_type,
             'expression': expression,
         })
+        self.save_expressions_file()
+        ui.message('Query saved')
+
+    def save_expressions_file(self):
         with open(SAVED_EXPRESSIONS_FILE_NAME, 'w', encoding='utf-8') as file:
             json.dump(self.saved_expressions, file, ensure_ascii=False, indent=2)
-        ui.message('Query saved')
 
     def set_output(self, output):
         self.output.SetValue(output if output is not None else '')
+
+    def get_auto_complete_choices(self):
+        return [
+            item['expression']
+            for item in self.saved_expressions
+            if item['type'] == self.expression_type
+        ]
 
     def exit_with_error(self, message):
         wx.CallAfter(