+++ addon/globalPlugins/nvda_json/json_query_dialog.py
@@ -25,9 +25,9 @@ class JsonQueryDialog(JsonManipulatorDialog):
     }
 
     def __init__(self, parent, text, multi):
-        super(JsonQueryDialog, self).__init__(parent, text, title = 'JSON Query')
         self.query_engine = JsonQueryDialog.query_engines[config.conf['json']['query_engine']]
         self.expression_type = self.query_engine['type']
+        super(JsonQueryDialog, self).__init__(parent, text, title = 'JSON Query')
         self.label_manipulation_expression.SetLabel(self.query_engine['edit_label'])
         self.multi = multi
         self.set_output(self.parse_text(self.text, self.multi))