+++ addon/globalPlugins/nvda_json/json_template_dialog.py
@@ -9,8 +9,8 @@ from .abstract import JsonManipulatorDialog
 
 class JsonTemplateDialog(JsonManipulatorDialog):
     def __init__(self, parent, text):
-        super(JsonTemplateDialog, self).__init__(parent, text, title = 'JSON template')
         self.expression_type = 'jsonpointer-template'
+        super(JsonTemplateDialog, self).__init__(parent, text, title = 'JSON template')
         self.label_manipulation_expression.SetLabel('Template')
         self.original_text.SetValue(text)